Remove extra stringent check in Forward, remove multi-sample
Check removed since that prevented fixing top level random nodes,
a frequently useful functionality. E.g. Doomsday with

--engine Forward
--model.z 1.0
--model.y NA
--model.rate 1.0

did not run. Downside is that this could miss fixing downstream nodes
but even that could be helpful in some scenarios..

Removed multi-sample so that can be fed to posterior inference easily